More Code Improvements and Bug Fixes

+++ src/main/java/bot/listeners/moderation/LinkControl.java
@@ -1,25 +1,26 @@
 package bot.listeners.moderation;
 
-import bot.Main;
 import bot.module.Module;
+import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
-import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import org.jetbrains.annotations.NotNull;
 
-import java.util.concurrent.TimeUnit;
+import java.time.Duration;
 
-public class LinkControl extends ListenerAdapter {
+public class LinkControl extends ModuleController {
+    private final Duration delay = Duration.ofSeconds(5);
 
-    public void onMessageReceived(@NotNull MessageReceivedEvent event) {
-        if (!Main.getModuleManager().isEnabled(Module.LINK_CONTROL)) return;
-        if (event.getAuthor().isBot()) return;
-        if (!event.isFromGuild()) return;
-        if (!event.getChannel().getName().contains("promotion")) {
-            String msg = event.getMessage().getContentRaw();
-            if (msg.contains("discord.gg/")) {
-                event.getMessage().delete().queue();
-                event.getChannel().sendMessage("No advertising in channels other than <#743858602997186612> please!").queue(m -> m.delete().queueAfter(5, TimeUnit.SECONDS));
-            }
+    public LinkControl() {
+        super(Module.LINK_CONTROL);
+    }
+
+    public void check(MessageReceivedEvent event) {
+        if (!event.getChannel().getName().contains("promotion") && !event.getMessage().getInvites().isEmpty()) {
+            event.getMessage().delete().queue();
+            event.getChannel()
+                .sendMessage("No advertising in channels other than <#743858602997186612> please!")
+                .flatMap(Message::delete)
+                .delay(delay) // delay the delete for 5 Seconds
+                .queue();
         }
     }
 }
